[CLEANUP] Final cleanup of TCA Overrides
Some of the files still miss the die statement in
the first line. To be consistent, add it to every
file.

One file had a check for TYPO3_MODE === BE which
is wrong in these files because the result is
cached.

If the first hit with empty cache is done in a FE
request, a wrong result would be cached.

Releases: 6.2, 6.3
Resolves: #60317
Change-Id: I839499100da28ba11d5dcbf81fde87b30b61de81
Reviewed-on: https://review.typo3.org/31621
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Markus Klein <klein.t3@mfc-linz.at>
Tested-by: Markus Klein <klein.t3@mfc-linz.at>