[BUGFIX] Fix return URL in OpenID wizard
The URL must be absolute and the parameters are cleaned up.
The returnURL sent to the OpenID provider must match the trusted
domain so it must be an absolute URL.

The code of the wizard is cleaned as well and also the closing curly
brace of a fluid translate viewhelper call is added.

Resolves: #66227
Releases: master
Change-Id: Ic745d5ea9ba6b1d949a65509ec8d0983c19b31f8
Reviewed-on: http://review.typo3.org/38415
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>