[BUGFIX] Correct normalization of openid url
Since the normalization of the openid is not working correctly, an openid
without a protocol and without a user in the database will never do a
request to the openid server.

So we use the normalization of the openid library build in. Then the
normalization is done right.

Resolves: #7
Related: #6
Releases: master